Move auth and encryption-related code into related mods
