+++ p2panda-spaces/src/auth/mod.rs
@@ -1,0 +1,1 @@
+pub mod orderer;
