--- p2panda-spaces/src/event.rs
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
-
-use crate::group::Group;
-use crate::space::Space;
-
-pub enum Event<S, F, M, C, RS> {
-    JoinedSpace(Space<S, F, M, C, RS>),
-    JoinedGroup(Group),
-    Message(M),
-}
