--- p2panda-spaces/src/traits.rs
@@ -1,1 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
